Delete moved files and fix build breaks
